+++ teensybdmdevice.cpp
@@ -27,9 +27,9 @@ void TeensyBDMDevice::reset(void)
 
 }
 
-uint32_t TeensyBDMDevice::getStatus(void)
+quint8 TeensyBDMDevice::getStatus(void)
 {
-    uint32_t status = 0;
+    quint8 status = 0;
 
     return status;
 }
@@ -58,104 +58,104 @@ void TeensyBDMDevice::assertTA(void)
 
 }
 
-uint8_t TeensyBDMDevice::read_mem_byte(uint32_t address)
+quint8 TeensyBDMDevice::read_mem_byte(quint32 address)
 {
-    uint8_t val = 0;
+    quint8 val = 0;
 
     return val;
 }
 
-uint16_t TeensyBDMDevice::read_mem_word(uint32_t address)
+quint16 TeensyBDMDevice::read_mem_word(quint32 address)
 {
-    uint16_t val = 0;
+    quint16 val = 0;
 
     return val;
 }
 
-uint32_t TeensyBDMDevice::read_mem_long(uint32_t address)
+quint32 TeensyBDMDevice::read_mem_long(quint32 address)
 {
-    uint32_t val = 0;
+    quint32 val = 0;
 
     return val;
 }
 
-void TeensyBDMDevice::read_memblock_byte(uint32_t address, QByteArray &arr)
+void TeensyBDMDevice::read_memblock_byte(quint32 address, QByteArray &arr)
 {
 
 }
 
-void TeensyBDMDevice::read_memblock_word(uint32_t address, QByteArray &arr)
+void TeensyBDMDevice::read_memblock_word(quint32 address, QByteArray &arr)
 {
 
 }
 
-void TeensyBDMDevice::read_memblock_long(uint32_t address, QByteArray &arr)
+void TeensyBDMDevice::read_memblock_long(quint32 address, QByteArray &arr)
 {
 
 }
 
-void TeensyBDMDevice::write_mem_byte(uint32_t address, uint8_t byte)
+void TeensyBDMDevice::write_mem_byte(quint32 address, quint8 byte)
 {
 
 }
 
-void TeensyBDMDevice::write_mem_word(uint32_t address, uint16_t word)
+void TeensyBDMDevice::write_mem_word(quint32 address, quint16 word)
 {
 
 }
 
-void TeensyBDMDevice::write_mem_long(uint32_t address, uint32_t longw)
+void TeensyBDMDevice::write_mem_long(quint32 address, quint32 longw)
 {
 
 }
 
-void TeensyBDMDevice::write_memblock_byte(uint32_t address, QByteArray &arr)
+void TeensyBDMDevice::write_memblock_byte(quint32 address, QByteArray &arr)
 {
 
 }
 
-void TeensyBDMDevice::write_memblock_word(uint32_t address, QByteArray &arr)
+void TeensyBDMDevice::write_memblock_word(quint32 address, QByteArray &arr)
 {
 
 }
 
-void TeensyBDMDevice::write_memblock_long(uint32_t address, QByteArray &arr)
+void TeensyBDMDevice::write_memblock_long(quint32 address, QByteArray &arr)
 {
 
 }
 
-uint32_t TeensyBDMDevice::read_reg(int regNum)
+quint32 TeensyBDMDevice::read_reg(int regNum)
 {
-    uint32_t regValue = 0;
+    quint32 regValue = 0;
 
     return regValue;
 }
 
-void TeensyBDMDevice::write_reg(int regnum, uint32_t value)
+void TeensyBDMDevice::write_reg(int regnum, quint32 value)
 {
 
 }
 
-uint32_t TeensyBDMDevice::read_creg(int regNum)
+quint32 TeensyBDMDevice::read_creg(int regNum)
 {
-    uint32_t regValue = 0;
+    quint32 regValue = 0;
 
     return regValue;
 }
 
-void TeensyBDMDevice::write_creg(int regnum, uint32_t value)
+void TeensyBDMDevice::write_creg(int regnum, quint32 value)
 {
 
 }
 
-uint32_t TeensyBDMDevice::read_dbgreg(int regnum)
+quint32 TeensyBDMDevice::read_dbgreg(int regnum)
 {
-    uint32_t regValue = 0;
+    quint32 regValue = 0;
 
     return regValue;
 }
 
-void TeensyBDMDevice::write_dbgreg(int regnum, uint32_t value)
+void TeensyBDMDevice::write_dbgreg(int regnum, quint32 value)
 {
 }
 
@@ -164,9 +164,9 @@ void TeensyBDMDevice::send_command(BDMCommand *command)
 
 }
 
-uint32_t TeensyBDMDevice::receive_result(void)
+quint32 TeensyBDMDevice::receive_result(void)
 {
-    uint32_t value = 0;
+    quint32 value = 0;
 
     return value;
 }
